# docs/swagger.yaml
definitions:
...
  getCurrentOrder.Response:
    properties:
      created_at:
        example: 2020-01-01 01:02:03 UTC
        type: string
      delivery_Address:
        example: 122 address
        type: string
      items:
        items:
          $ref: '#/definitions/getCurrentOrder.item'
        type: array
      restaurant_Address:
        example: 123 address
        type: string
      restaurant_Name:
        example: Mac
        type: string
      restaurant_Phone:
        example: "89056663333"
        type: string
      reward:
        example: 12
        type: number
      status:
        example: pending
        type: string
    type: object
  getCurrentOrder.item:
    properties:
      item_name:
        example: Burger with cheese
        type: string
      quantity:
        example: 3
        type: integer
    type: object
...
  getOrderByID.Response:
    properties:
      courierName:
        example: John
        type: string
      created_at:
        example: "2020-09-20T14:14:15+09:00"
        type: string
      delivery_Address:
        example: 1222 address
        type: string
      items:
        items:
          $ref: '#/definitions/getOrderByID.item'
        type: array
      restaurant_Address:
        example: 123 address
        type: string
      restaurant_Name:
        example: Mac
        type: string
      restaurant_Phone:
        example: "89055463333"
        type: string
      status:
        example: pending
        type: string
      total_price:
        example: 300
        type: number
      user_name:
        example: Bill
        type: string
    type: object
  getOrderByID.item:
    properties:
      item_name:
        example: burger
        type: string
      price:
        example: 60
        type: number
      quantity:
        example: 5
        type: integer
    type: object
  getOrdersForUser.Response:
    properties:
      ordersStruct:
        items:
          $ref: '#/definitions/ordersStruct.Order'
        type: array
    type: object
  getPendingOrders.Response:
    properties:
      pending_orders:
        items:
          $ref: '#/definitions/ordersStruct.OrderForCourier'
        type: array
    type: object
...
  orderAssign.Response:
    properties:
      courierName:
        example: Bill
        type: string
      created_at:
        example: "2020-01-01T00:00:00+09:00"
        type: string
      delivery_Address:
        example: 1222 address
        type: string
      items:
        items:
          $ref: '#/definitions/orderAssign.item'
        type: array
      restaurant_Address:
        example: 123 address
        type: string
      restaurant_Name:
        example: Mac
        type: string
      restaurant_Phone:
        example: "89056666666"
        type: string
      status:
        example: pending
        type: string
      total_price:
        example: 300
        type: number
      user_name:
        example: Ivan
        type: string
    type: object
  orderAssign.item:
    properties:
      item_name:
        example: Burger with cheese
        type: string
      price:
        example: 100
        type: number
      quantity:
        example: 3
        type: integer
    type: object
...
  ordersStruct.Item:
    properties:
      item_name:
        example: Burger with cheese
        type: string
      item_price:
        example: 100
        type: number
      menu_item_id:
        example: 1
        type: integer
      quantity:
        example: 3
        type: integer
    type: object
  ordersStruct.Order:
    properties:
      created_at:
        example: "2013-08-20T18:08:41+00:00"
        type: string
      delivery_Address:
        example: 1223 address
        type: string
      items:
        items:
          $ref: '#/definitions/ordersStruct.Item'
        type: array
      restaurant_Address:
        example: 123 address
        type: string
      restaurant_Name:
        example: Mac
        type: string
      restaurant_Phone:
        example: "89056666666"
        type: string
      status:
        example: pending
        type: string
      total_price:
        example: 300
        type: number
      user_name:
        example: Ivan
        type: string
    type: object
  ordersStruct.OrderForCourier:
    properties:
      created_at:
        example: "2013-08-20T18:08:41+00:00"
        type: string
      delivery_Address:
        example: 1223 address
        type: string
      items:
        items:
          $ref: '#/definitions/ordersStruct.Item'
        type: array
      order_id:
        example: 1
        type: integer
      restaurant_Address:
        example: 123 address
        type: string
      restaurant_Name:
        example: Mac
        type: string
      restaurant_Phone:
        example: "89056666666"
        type: string
      reward:
        example: 30
        type: number
      user_phone:
        example: Ivan
        type: string
    type: object
  placeorder.Request:
    properties:
      address:
        example: 123 address
        type: string
      items:
        items:
          properties:
            menuitem_id:
              example: 6
              type: integer
            quantity:
              example: 5
              type: integer
          type: object
        type: array
      restaurant_id:
        example: 14
        type: integer
    type: object
  placeorder.Response:
    properties:
      created_at:
        example: Tue, 17 Jun 2025 00:25:16 +0000
        type: string
      items:
        items:
          properties:
            menuitem_id:
              example: 6
              type: integer
            quantity:
              example: 5
              type: integer
          type: object
        type: array
      order_id:
        example: 12
        type: integer
      restaurant_id:
        example: 14
        type: integer
      status:
        example: pending
        type: string
      user_address:
        example: 123 address
        type: string
    type: object
...
  response.Response:
    properties:
      error:
        example: error message
        type: string
    type: object
info:
  contact: {}
  description: REST API for food delivery
  title: GodFood API
  version: "1.0"
paths:
...
  /orders:
    get:
      consumes:
      - application/json
      description: Возвращает полную информацию по заказам для авторизованного пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Заказ успешно получен
          schema:
            $ref: '#/definitions/getOrdersForUser.Response'
        "204":
          description: Нет заказов
        "400":
          description: Неккоректные данные
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Доступ Запрещен
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Получение заказов по JWT
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Создает новый заказ
      parameters:
      - description: Данные для добавления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/placeorder.Request'
      produces:
      - application/json
      responses:
        "200":
          description: 'Новый заказ успешно создан '
          schema:
            $ref: '#/definitions/placeorder.Response'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Серверная Ошибка
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Создание нового заказа авторизованным пользователем
      tags:
      - Orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает полную информацию по заказу(если авторизованный пользователь
        им владеет)
      parameters:
      - description: ID Заказа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Заказ успешно получен
          schema:
            $ref: '#/definitions/getOrderByID.Response'
        "400":
          description: Неккоректные данные
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Доступ Запрещен
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Получение заказа по айди
      tags:
      - Orders
  /orders/{id}/assign:
    patch:
      consumes:
      - application/json
      description: Назначает заказ на курьера
      parameters:
      - description: ID Заказа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Заказ назначен
          schema:
            $ref: '#/definitions/orderAssign.Response'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Заказ не найден
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Серверная Ошибка
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Взятие заказа курьером
      tags:
      - Orders
  /orders/current:
    get:
      consumes:
      - application/json
      description: Возвращает полную информацию по заказу, который везет авторизованный
        курьер
      produces:
      - application/json
      responses:
        "200":
          description: Заказ успешно получен
          schema:
            $ref: '#/definitions/getCurrentOrder.Response'
        "403":
          description: Неавторизован
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Заказ не найден
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Получение нынешнего заказа курьера
      tags:
      - Orders
...
  /orders/pending:
    get:
      consumes:
      - application/json
      description: Возвращает все заказы, которые еще не взяты
      produces:
      - application/json
      responses:
        "200":
          description: Заказы успешно получен
          schema:
            $ref: '#/definitions/getPendingOrders.Response'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Заказы не найдены
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Получение всех доступных для доставки заказов
      tags:
      - Orders
...
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"